Introduced context and setup peer sign on
Create a context to compose essential instances/data required downstream of the feta assets/apps. Set upo sign on such that when a peer starts uo their server they get (not yet connected) to a websocket host and record which host they are connected to so other peers can find them.

Also created scripts and docker compose setup to help with setup of dev environment locally